requirements.txt
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
anthropic>=0.39.0
requests>=2.31.0
python-dotenv>=1.0.0
